fix grayscale conversion. Make DRY. add img result
fix grayscale conversion: in this function, the image was read in using mpimg, not cv2
  change grayscale param from cv2.COLOR_BGR2GRAY to cv2.COLOR_RGB2GRAY

Prior, the function to compute both sobelx and sobely, choosing which to return at end.
Now function chooses to compute only sobelx or sobely, at the beginning of the func.

prior, I did not use the thresh_min and thresh_max that were passed in

Added notes, and final output.