branch for debugging plot_8_gray_images(), L_28
goal was to group all the code for the final 8 plots together into one cell,
so that those 8 plots can be seen side-by-side, rather than being split by code.

So I moved and updated code for creating those 8 images.
And wrote plot_8_gray_images() to accept all 8.
But there is some but having something to do with the subplot params.

Hence, I moved this version to a new testing "try" (testing) branch to
work out the bug (if ever).  Two cells can then be moved back into
L_29_hls_color_threshold.ipynb, and a bunch of other cells can be deleted,
once the bug has been worked out.

Note: in debugging, I now have plot_8_gray_images in TWO cells:
The second def is the cell I'm currently working in.
Point was to have all the code I needed to run for debugging in 1 place,
to avoid scrolling up and down to re-run the two relevant cells.

If I do come back to debug this file, it would be a good idea to
REMOVE all the irrelevant cells from this notebook !!

This file is also used to temp test other things, such as LaTeX commands.
So don't rely on other cells being correct, or relevant !